+++ lib/repeatable/expression/daily.rb
@@ -1,0 +1,11 @@
+# typed: strict
+module Repeatable
+  module Expression
+    class Daily < Date
+      sig { override.params(_date: ::Date).returns(T::Boolean) }
+      def include?(_date)
+        true
+      end
+    end
+  end
+end